Added a fresh copy of the default styleguidekit assets directory and amended patternlab-config.
In order to display a custom front end for Pattern Lab, you must tell the config where the static assets for the front end are. I've copied the default assets from Pattern Lab's node_modules directory into the root of the project, so that these files can be amended separately to the default ones, which are ignored by git.

The project will use the default mustache templating engine files which are stored in node_modules in order to generate the static public files. The assets simply describe how the front end looks.