Deleted all Pattern Lab files and updated README.
This issue relates to the decision to remove PL from the repo and instead use a custom made styleguide generator that is currently being worked on but is in a state to use and develop for production.

This commit removes the core PL files and updates the README to reflect that.