Significant change to gulp createNav, read on for more.
Previous to this commit, the createNav function was buidling a single level UL with LIs for each .html file found anywhere in the public directory. I didn't foresee that this would cause problems when it comes to nesting ULs inside of LIs in order to clearly create a navigation and sub-navigation dropdowns.

The update, although small, took a while to figure out and essentially loops over all the returned items and creates a multi-level menu that matches the directory structure. This means we end up with top-level folders, as well as sub folders and then links to pages.

This is tested and seems to be working as expected now, though the code could really benefit from some tidying up, sanitisation and abstracting. I will get to this as soon as I can.