Added JS methods to obscure password fields that have been made visible after a timeout.
Password field text that has been made visible has been made more difficult to read by lowering the text opacity when the input filled class is applied and the input field is text.

If the visibility button is clicked and then no further input in the password field is detected, the fields type will be set back to password after 3 seconds.

If the user continues to type into the field after making the password visible, the countdown is stopped.

If the user has typed into the field and stopped the countdown, then navigates their focus away from the field, it will obscure the password after 1.5 seconds.